build fixed

+++ common/azimuth_elevation.cpp
@@ -1,15 +1,17 @@
 #include "azimuth_elevation.h"
 
+#include "vec.h"
+
 #include <cmath>
 
-Vec3f sphere(float cosz, float sinz, float theta) {
+Vec3f sphere(double cosz, double sinz, double theta) {
     return Vec3f(cosz * cos(theta), sinz, cosz * sin(theta));
 }
 
-Vec3f point_on_sphere(float az, float el) {
+Vec3f point_on_sphere(double az, double el) {
     return sphere(cos(el), sin(el), az);
 }
 
-Vec3f sphere_point(float z, float theta) {
+Vec3f sphere_point(double z, double theta) {
     return sphere(std::sqrt(1 - z * z), z, theta);
 }
